Performance update
added dialog when delete post
added more lazy pages
reducing bundle size
UI improvements
Backend Performance Update
dependences spilted into dev and production
ready for production

+++ backend/users/migrations/0017_alter_socialuser_phone.py
@@ -13,7 +13,7 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='socialuser',
             name='phone',
-            field=models.IntegerField(default=101010101010, unique=True),
+            field=models.BigIntegerField(default=101010101010, unique=True),
             preserve_default=False,
         ),
     ]
